Make the software more robust by moving connected parts together
There is a strong connection between the filter that we apply on the
messages (e.g. ALL, REJECTED, etc.) and the content of the
messageIdToStatus map (e.g. for REJECTED messages we only put the REJECTED
messages to the messageIdToStatus map). Sometimes, when we try to make a
decision based on the filter, we fall back to one of the possible values
(e.g. if we don't know what the filter is -- please note, it is an
internal error --, we fall back to ALL), but we don't modify the content
of the messageIdToStatus map. It can result a wrong message list. It is
better to move the two things (the query of the messageIdToStatus map, and
the decision that I described) to a common place (to the FirebaseUtils)
which makes it more difficult to make a mistake in this regard.