Workaround the bug that Firebase returns null for displayName
Firebase registers (email registration) the user in two steps:
  1. Registration with email + password
  2. Change the displayname of the user
Server side triggers are triggered after step 1 when the display name may
or may not be available. Because of this the name cannot be stored to the
Firebase online DB on server side (easily / I don't know how to do that)

At the first profile store we store the default name, or the user can
change the name if they want.