Restrict the user with the child date
The app is currently not prepared to handle if the user child was born
many many days ago (it would be bad to start with 100s of messages) or
when the child has not been born. This is a temporary fix for this issue
until the introduction of batch messages and minus day messages.